Expose Cursor for raw statement row access
Allows for direct access to a column value as any SQLite type without
the intermediary array of optional values.

    let stmt = db.prepare("SELECT age FROM users")
    while stmt.step() {
        let age = stmt.cursor[0] as String // "30"
        // ...
    }

The cursor should be the default value that is yielded to the for loop,
but Swift's current inability to resolve the following prevents that
from working well:

    for row in stmt {
        let age = row[0] as Binding? as? String // passes compilation
        let age = row[0] as? String             // fails
    }

If the above case is resolved (rdar://19997098), then we can avoid
automatically unpacking every row into an array by default. This means
the following won't work:

    let rows = Array(stmt)

But the following will:

    let rows = map(stmt) { Array($0) }

Signed-off-by: Stephen Celis <stephen@stephencelis.com>